lvl3

+++ lab-socket-hunt/treasure_hunter.c
@@ -224,7 +224,7 @@ int main(int argc, char *argv[]) {
 				}
 				break;
 			case 3:;
-				int m = ntohl(par);
+				int m = (int)ntohl(par);
 				printf("%d",m);
 				unsigned int sum = 0;
 				for (int j = 0; j < m; j++){
